requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Deprecated==1.2.15
idna==3.10
pycparser==2.22
PyGithub==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.0
